cleanup(notes): Cleaned up note types
