+++ src/api-calls/game-log/_game-log.type.ts
@@ -1,0 +1,6 @@
+import { Timestamp } from "firebase/firestore";
+import { Roll } from "types/DieRolls.type";
+
+export type GameLogDocument = Omit<Roll, "timestamp"> & {
+  timestamp: Timestamp;
+};
