+++ src/api-calls/game-log/_getRef.ts
@@ -7,6 +7,7 @@ import {
   doc,
 } from "firebase/firestore";
 import { Roll } from "types/DieRolls.type";
+import { GameLogDocument } from "./_game-log.type";
 
 export function constructCampaignGameLogCollectionPath(campaignId: string) {
   return `/campaigns/${campaignId}/game-log`;
@@ -23,14 +24,14 @@ export function getCampaignGameLogCollection(campaignId: string) {
   return collection(
     firestore,
     constructCampaignGameLogCollectionPath(campaignId)
-  ) as CollectionReference<Roll>;
+  ) as CollectionReference<GameLogDocument>;
 }
 
 export function getCampaignGameLogDocument(campaignId: string, logId: string) {
   return doc(
     firestore,
     constructCampaignGameLogDocPath(campaignId, logId)
-  ) as DocumentReference<Roll>;
+  ) as DocumentReference<GameLogDocument>;
 }
 
 export function constructCharacterGameLogCollectionPath(characterId: string) {
@@ -48,7 +49,7 @@ export function getCharacterGameLogCollection(characterId: string) {
   return collection(
     firestore,
     constructCharacterGameLogCollectionPath(characterId)
-  ) as CollectionReference<Roll>;
+  ) as CollectionReference<GameLogDocument>;
 }
 
 export function getCharacterGameLogDocument(
@@ -58,16 +59,19 @@ export function getCharacterGameLogDocument(
   return doc(
     firestore,
     constructCharacterGameLogDocPath(characterId, logId)
-  ) as DocumentReference<Roll>;
+  ) as DocumentReference<GameLogDocument>;
 }
 
-export type DatabaseLog = Omit<Roll, "timestamp"> & {
-  timestamp: Timestamp;
-};
-
-export function convertFromDatabase(log: DatabaseLog): Roll {
+export function convertFromDatabase(log: GameLogDocument): Roll {
   return {
     ...log,
     timestamp: log.timestamp.toDate(),
   } as Roll;
+}
+
+export function convertRollToGameLogDocument(roll: Roll): GameLogDocument {
+  return {
+    ...roll,
+    timestamp: Timestamp.fromDate(roll.timestamp),
+  };
 }
